requirements.txt
asgiref==3.2.10
autopep8==1.5.7
certifi==2021.5.30
chardet==4.0.0
Django==3.1
django-admin-thumbnails==0.2.5
django-ckeditor==6.1.0
django-countries==7.2.1
django-crispy-forms==1.12.0
django-js-asset==1.2.2
django-mptt==0.12.0
django-taggit==1.5.0
gunicorn==20.1.0
idna==2.10
Pillow==8.2.0
psycopg2-binary==2.8.6
pycodestyle==2.7.0
pytz==2021.1
requests==2.25.1
six==1.16.0
sqlparse==0.4.1
toml==0.10.2
urllib3==1.26.5
